+++ src/domain/meetups/entities/like.entity.ts
@@ -9,7 +9,7 @@ export class Like {
   @PrimaryColumn({ type: 'int', unsigned: true })
   public userId: number;
 
-  @PrimaryColumn({ type: 'bigint', unsigned: true })
+  @PrimaryColumn({ type: 'int', unsigned: true })
   public meetupId: number;
 
   @Column({ length: 16, nullable: true })
